requirements.txt
Flask
tensorflow
flask-cors
Pillow